C++ User created stack data structure w/ class
Simple class I made that implements a stack data structure.
This stack will push integer data to the stack, inserting at the top.
The viewStack() function will return the top value, and pop that previous top, iterating to the new location top points to upon deletion.
The stack is popped until no data is remaining. If the user forgets to pop the data, the automatic destructor will delete the stack in memory.